requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autoflake==2.2.0

billiard==4.1.0
black==23.3.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-yasg==1.21.6
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.8.0
prompt-toolkit==3.0.39
pyflakes==3.0.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
gunicorn
uvicorn
mysqlclient==2.1.1